requirements.txt
django
djangorestframework
python-dotenv
psycopg2-binary
Pillow
django-cors-headers
firebase-admin
boto3
django-storages
drf-spectacular
channels
stripe
gunicorn
uvicorn
fcm-django
firebase-admin
django_serverside_datatable